refactoring

+++ src/CatalogService/CatalogService.Api/Contracts/SearchProductsRequest.cs
@@ -7,8 +7,8 @@ public class SearchProductsRequest
 {
     public string? Text { get; init; }
     public Guid? CategoryId { get; init; }
-    public decimal? MinPrice { get; init; }
-    public decimal? MaxPrice { get; init; }
+    public long? MinPrice { get; init; }
+    public long? MaxPrice { get; init; }
 
     public ProductSortBy SortBy { get; init; } = ProductSortBy.Name;
     public SortDirection SortDir { get; init; } = SortDirection.Asc;
