+++ src/CatalogService/CatalogService.Application/Interfaces/IProductIntegrationEventMapper.cs
@@ -4,6 +4,6 @@ namespace CatalogService.Application.Interfaces;
 
 public interface IProductIntegrationEventMapper
 {
-    ProductPriceChangedEvent MapProductPriceChangedEvent(Guid productId, decimal oldPriceCents, decimal newPriceCents);
-    ProductCreatedEvent MapProductCreatedEvent(Guid productId, string name, string description, int priceCents, int quantityAvailable);
+    ProductPriceChangedEvent MapProductPriceChangedEvent(Guid productId, long oldPriceCents, long newPriceCents);
+    ProductCreatedEvent MapProductCreatedEvent(Guid productId, string name, string description, long priceCents, int quantityAvailable);
 }
