+++ src/CatalogService/CatalogService.Application/Models/ProductView.cs
@@ -1,3 +1,3 @@
 ﻿namespace CatalogService.Application.Models;
 
-public record ProductView(Guid Id, string Name, string? Description, decimal PriceCents, string Currency);
+public record ProductView(Guid Id, string Name, string? Description, long PriceCents, string Currency);
