+++ src/CatalogService/CatalogService.Application/Models/UpdateProductModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace CatalogService.Application.Models;
 
-public record UpdateProductModel(string? Name, string? Description, decimal? PriceCents, string? Currency, bool? IsActive);
+public record UpdateProductModel(string? Name, string? Description, long? PriceCents, string? Currency, bool? IsActive);
