+++ src/CatalogService/CatalogService.Application/Queries/SearchProducts/SearchProductsQuery.cs
@@ -9,8 +9,8 @@ namespace CatalogService.Application.Queries.SearchProducts;
 public record SearchProductsQuery(
     string? Text,
     Guid? CategoryId,
-    decimal? MinPrice,
-    decimal? MaxPrice,
+    long? MinPrice,
+    long? MaxPrice,
     ProductSortBy SortBy,
     SortDirection SortDirection,
     int Page,
