+++ src/OrderService/OrderService.Application/Models/OrderView.cs
@@ -15,7 +15,7 @@ public record OrderView
     public string Description { get; set; } = string.Empty;
     public int WeightGrams { get; set; }
     public int Status { get; set; }
-    public decimal CostCents { get; set; }
+    public long CostCents { get; set; }
     public string? CourierNote { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime? AssignedAt { get; set; }
